+++ deployer/drupal/config/set.php
@@ -40,3 +40,7 @@ set('buffer_config', function () {
 });
 
 set('feature_index_app_type', 'drupal');
+
+set('drush', function () {
+  return which('drush');
+});